+++ ttrs_field.c
@@ -27,8 +27,8 @@ bool can_place_in_field(const t_field_ptr field_ptr,
 void place_mino_on_field(t_field_ptr current_field, t_mino *mino)
 {
 	t_point pos = mino->pos;
-	for (size_t i = 0; i < mino->mino_type.width; i++) {
-		for (size_t j = 0; j < mino->mino_type.width; j++) {
+	for (int i = 0; i < mino->mino_type.width; i++) {
+		for (int j = 0; j < mino->mino_type.width; j++) {
 			if (mino->mino_type.shape[i][j]) {
 				current_field[pos.row + i][pos.col + j] = mino->mino_type.shape[i][j];
 			}
@@ -46,17 +46,17 @@ static void erase_filled_line(t_field_ptr field_ptr, int row)
 
 static bool is_filled_line(t_field_line_ptr line)
 {
-	size_t filled_cell = 0;
-	for (size_t i = 0; i < FIELD_COL; i++) {
+	int filled_cell = 0;
+	for (int i = 0; i < FIELD_COL; i++) {
 		filled_cell += line[i];
 	}
 	return filled_cell == FIELD_COL;
 }
 
-size_t erase_filled_lines(t_field_ptr field_ptr)
+int erase_filled_lines(t_field_ptr field_ptr)
 {
-	size_t count = 0;
-	for (size_t i = 0; i < FIELD_ROW; i++) {
+	int count = 0;
+	for (int i = 0; i < FIELD_ROW; i++) {
 		if (is_filled_line(field_ptr[i])) {
 			count++;
 			erase_filled_line(field_ptr, i);
